Move flask_app (where layout is defined) to LayoutT1D
