+++ run_flask_app.py
@@ -4,7 +4,7 @@ from app import app
 import sys
 sys.path.insert(0, "../")
 
-from dashWebApplication.flask_app import layout
+from dashWebApplication.LayoutT1D import layout
 
 # This is apparently okay to deploy too.
 if __name__ == '__main__':
